Extract case endpoints
