+++ cg/server/api.py
@@ -6,7 +6,6 @@ from pathlib import Path
 from typing import Any
 
 from flask import Blueprint, abort, g, jsonify, make_response, request
-from google.oauth2 import id_token
 from pydantic.v1 import ValidationError
 from requests.exceptions import HTTPError
 from sqlalchemy.exc import IntegrityError
@@ -18,11 +17,9 @@ from cg.apps.orderform.json_orderform_parser import JsonOrderformParser
 from cg.constants import ANALYSIS_SOURCES, METAGENOME_SOURCES
 from cg.constants.constants import FileFormat
 from cg.exc import (
-    CaseNotFoundError,
     OrderError,
     OrderExistsError,
     OrderFormError,
-    OrderMismatchError,
     OrderNotDeliverableError,
     OrderNotFoundError,
     TicketCreationError,
@@ -32,7 +29,6 @@ from cg.meta.orders import OrdersAPI
 from cg.meta.orders.ticket_handler import TicketHandler
 from cg.models.orders.order import OrderIn, OrderType
 from cg.models.orders.orderform_schema import Orderform
-from cg.server.dto.delivery_message.delivery_message_request import DeliveryMessageRequest
 from cg.server.dto.delivery_message.delivery_message_response import DeliveryMessageResponse
 from cg.server.dto.orders.order_delivery_update_request import OrderDeliveredUpdateRequest
 from cg.server.dto.orders.order_patch_request import OrderDeliveredPatch
@@ -52,7 +48,6 @@ from cg.store.models import (
     Analysis,
     Application,
     ApplicationLimitations,
-    Case,
     Customer,
     IlluminaSampleSequencingMetrics,
     Pool,
@@ -123,101 +118,6 @@ def submit_order(order_type):
 
     if error_message:
         return abort(make_response(jsonify(message=error_message), http_error_response))
-
-
-@BLUEPRINT.route("/cases")
-def get_cases():
-    """Return cases with links for a customer from the database."""
-    enquiry: str = request.args.get("enquiry")
-    action: str = request.args.get("action")
-
-    customers: list[Customer] = _get_current_customers()
-    cases: list[Case] = _get_cases(enquiry=enquiry, action=action, customers=customers)
-
-    nr_cases: int = len(cases)
-    cases_with_links: list[dict] = [case.to_dict(links=True) for case in cases]
-    return jsonify(families=cases_with_links, total=nr_cases)
-
-
-def _get_current_customers() -> list[Customer] | None:
-    """Return customers if the current user is not an admin."""
-    return g.current_user.customers if not g.current_user.is_admin else None
-
-
-def _get_cases(
-    enquiry: str | None, action: str | None, customers: list[Customer] | None
-) -> list[Case]:
-    """Get cases based on the provided filters."""
-    return db.get_cases_by_customers_action_and_case_search(
-        case_search=enquiry,
-        customers=customers,
-        action=action,
-    )
-
-
-@BLUEPRINT.route("/cases/<case_id>")
-def parse_case(case_id):
-    """Return a case with links."""
-    case: Case = db.get_case_by_internal_id(internal_id=case_id)
-    if case is None:
-        return abort(HTTPStatus.NOT_FOUND)
-    if not g.current_user.is_admin and (case.customer not in g.current_user.customers):
-        return abort(HTTPStatus.FORBIDDEN)
-    return jsonify(**case.to_dict(links=True, analyses=True))
-
-
-@BLUEPRINT.route("/cases/delivery_message", methods=["GET"])
-def get_cases_delivery_message():
-    delivery_message_request = DeliveryMessageRequest.model_validate(request.args)
-    try:
-        response: DeliveryMessageResponse = delivery_message_service.get_cases_message(
-            delivery_message_request
-        )
-        return jsonify(response.model_dump()), HTTPStatus.OK
-    except (CaseNotFoundError, OrderMismatchError) as error:
-        return jsonify({"error": str(error)}), HTTPStatus.BAD_REQUEST
-
-
-@BLUEPRINT.route("/cases/<case_id>/delivery_message", methods=["GET"])
-def get_case_delivery_message(case_id: str):
-    delivery_message_request = DeliveryMessageRequest(case_ids=[case_id])
-    try:
-        response: DeliveryMessageResponse = delivery_message_service.get_cases_message(
-            delivery_message_request
-        )
-        return jsonify(response.model_dump()), HTTPStatus.OK
-    except CaseNotFoundError as error:
-        return jsonify({"error": str(error)}), HTTPStatus.BAD_REQUEST
-
-
-@BLUEPRINT.route("/families_in_collaboration")
-def parse_families_in_collaboration():
-    """Return cases in collaboration."""
-
-    customer_internal_id = request.args.get("customer")
-    workflow = request.args.get("data_analysis")
-    case_search_pattern = request.args.get("enquiry")
-
-    customer = db.get_customer_by_internal_id(customer_internal_id=customer_internal_id)
-
-    cases = db.get_cases_by_customer_workflow_and_case_search(
-        customer=customer, workflow=workflow, case_search=case_search_pattern
-    )
-
-    case_dicts = [case.to_dict(links=True) for case in cases]
-    return jsonify(families=case_dicts, total=len(cases))
-
-
-@BLUEPRINT.route("/families_in_collaboration/<family_id>")
-def parse_family_in_collaboration(family_id):
-    """Return a family with links."""
-    case: Case = db.get_case_by_internal_id(internal_id=family_id)
-    customer: Customer = db.get_customer_by_internal_id(
-        customer_internal_id=request.args.get("customer")
-    )
-    if case.customer not in customer.collaborators:
-        return abort(HTTPStatus.FORBIDDEN)
-    return jsonify(**case.to_dict(links=True, analyses=True))
 
 
 @BLUEPRINT.route("/samples")
