+++ cg/server/app.py
@@ -8,6 +8,7 @@ from sqlalchemy.orm import scoped_session
 
 from cg.server import admin, api, ext, invoices
 from cg.server.app_config import app_config
+from cg.server.endpoints.cases import CASES_BLUEPRINT
 from cg.server.endpoints.samples import SAMPLES_BLUEPRINT
 from cg.store.database import get_scoped_session_registry
 from cg.store.models import (
@@ -88,6 +89,7 @@ def _register_blueprints(app: Flask):
     app.register_blueprint(invoices.BLUEPRINT, url_prefix="/invoices")
     app.register_blueprint(oauth_bp, url_prefix="/login")
     app.register_blueprint(SAMPLES_BLUEPRINT)
+    app.register_blueprint(CASES_BLUEPRINT)
     _register_admin_views()
 
     ext.csrf.exempt(api.BLUEPRINT)  # Protected with Auth header already
